Fix heartbeating
